Implement order, invoice and shipment import logic from Exact
The data presented by Exact can be imported and fetched now using
the REST API calls. The provided data from an Exact order is
migrated to an actual Magento order and all the data is set propperly.

The other way around is that a Magento order is normalized into the
expected format for Xcore/Exact to read all the required data again.